# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Sona API
  version: 0.5.0
servers:
  - url: http://sosee.eu.cc:6699/api/v1
paths:
...
  /me/playlists/{playlistId}/tracks/{trackId}:
    put:
      operationId: addPlaylistTrack
      parameters:
        - $ref: "#/components/parameters/PlaylistId"
        - $ref: "#/components/parameters/TrackId"
      responses:
        "204":
          description: 曲目已加入歌单
    delete:
      operationId: removePlaylistTrack
      parameters:
        - $ref: "#/components/parameters/PlaylistId"
        - $ref: "#/components/parameters/TrackId"
      responses:
        "204":
          description: 曲目已移出歌单
...
  /me/tracks/{trackId}:
    delete:
      operationId: hideTrack
      parameters:
        - $ref: "#/components/parameters/TrackId"
      responses:
        "204":
          description: 歌曲已加入当前用户垃圾桶
...
  /tracks:
    get:
      operationId: listTracks
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: childMode
          in: query
          schema: {type: boolean, default: false}
        - name: sort
          in: query
          schema: {type: string, enum: [TITLE, ARTIST, ALBUM, NEWEST], default: TITLE}
        - name: genre
          in: query
          schema: {type: string}
        - name: codec
          in: query
          schema: {type: string}
        - name: metadataStatus
          in: query
          schema: {type: string, enum: [LOCAL, SCRAPED, NEEDS_REVIEW, MANUAL]}
      responses:
        "200":
          description: 曲目列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /tracks/random:
    get:
      operationId: randomTracks
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: childMode
          in: query
          schema: {type: boolean, default: false}
      responses:
        "200":
          description: 按用户维护随机覆盖轮次；未抓取歌曲优先并受全局完播率加权，保证正常池最终全覆盖且批内不重复
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Track"
  /tracks/discovery:
    get:
      operationId: discoveryTracks
      parameters:
        - name: limit
          in: query
          schema: {type: integer, default: 10}
        - name: childMode
          in: query
          schema: {type: boolean, default: false}
      responses:
        "200":
          description: 发现池试听歌曲
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Track"
  /tracks/{trackId}:
    get:
      operationId: getTrack
      parameters:
        - $ref: "#/components/parameters/TrackId"
      responses:
        "200":
          description: 曲目详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "404":
          $ref: "#/components/responses/NotFound"
  /tracks/{trackId}/stream:
    get:
      operationId: streamTrack
      parameters:
        - $ref: "#/components/parameters/TrackId"
        - name: Range
          in: header
          schema:
            type: string
      responses:
        "200":
          description: 完整音频
          content:
            audio/*:
              schema:
                type: string
                contentMediaType: application/octet-stream
        "206":
          description: 音频字节范围
          headers:
            Accept-Ranges:
              schema:
                type: string
            Content-Range:
              schema:
                type: string
        "404":
          $ref: "#/components/responses/NotFound"
  /tracks/{trackId}/lyrics:
    get:
      operationId: getLyrics
      parameters:
        - $ref: "#/components/parameters/TrackId"
      responses:
        "200":
          description: 歌词
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lyrics"
        "404":
          $ref: "#/components/responses/NotFound"
  /tracks/{trackId}/artwork:
    get:
      operationId: getArtwork
      parameters:
        - $ref: "#/components/parameters/TrackId"
      responses:
        "200":
          description: 封面图片
          content:
            image/jpeg:
              schema:
                type: string
                contentMediaType: image/jpeg
            image/png:
              schema:
                type: string
                contentMediaType: image/png
        "404":
          $ref: "#/components/responses/NotFound"
...
  /library/tracks:
    get:
      operationId: managedTracks
      parameters:
        - name: poolType
          in: query
          schema: {type: string, enum: [PENDING, NORMAL, DISCOVERY]}
      responses:
        "200":
          description: 管理员歌曲列表
          content:
            application/json:
              schema: {type: array, items: {$ref: "#/components/schemas/Track"}}
  /library/tracks/{trackId}:
    patch:
      operationId: classifyTrack
      parameters:
        - $ref: "#/components/parameters/TrackId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [poolType, audienceType]
              properties:
                poolType: {type: string, enum: [PENDING, NORMAL, DISCOVERY]}
                audienceType: {type: string, enum: [GENERAL, CHILD]}
                genre: {type: string, maxLength: 40}
                region: {type: string, enum: [CN, KR, US, JP, OTHER]}
      responses:
        "200": {description: 歌曲池、儿童分类、曲风与地区已更新}
    delete:
      operationId: permanentlyDeleteTrack
      parameters:
        - $ref: "#/components/parameters/TrackId"
      responses:
        "204": {description: 音频文件及关联数据已删除}
  /library/tracks/{trackId}/metadata:
    patch:
      operationId: editTrackMetadata
      description: 仅修改 Sona 数据库，不回写原始音频文件
      parameters:
        - $ref: "#/components/parameters/TrackId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, artist, album, genre]
              properties:
                title: {type: string, maxLength: 200}
                artist: {type: string, maxLength: 200}
                album: {type: string, maxLength: 200}
                trackNumber: {type: [integer, "null"], minimum: 0}
                genre: {type: string, maxLength: 40}
      responses:
        "200":
          description: 数据库元数据已更新
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Track"}
  /library/tracks/{trackId}/rescrape:
    post:
      operationId: rescrapeTrack
      parameters:
        - $ref: "#/components/parameters/TrackId"
      responses:
        "202":
          description: 已清除人工锁定并开始重新扫描
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ScanStatus"}
  /library/tracks/upload:
    post:
      operationId: uploadTracks
      description: 上传 App 本地音频，扫描刮削后直接进入正常歌曲池（仅管理员）
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files: {type: array, items: {type: string, format: binary}}
      responses:
        "202": {description: 文件已上传并开始扫描刮削}
...
components:
  parameters:
    TrackId:
      name: trackId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PlaylistId:
      name: playlistId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: 未登录或会话失效
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    NotFound:
      description: 资源不存在
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    Track:
      type: object
      required: [id, title, artist, album, durationMs, codec, fileExtension, streamURL, hasLyrics, metadataStatus, poolType, audienceType, genre, region, artists]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        trackNumber:
          type: integer
        durationMs:
          type: integer
          format: int64
        codec:
          type: string
        fileExtension:
          type: string
        sampleRate:
          type: integer
        bitDepth:
          type: integer
        artworkURL:
          type: string
          format: uri-reference
        streamURL:
          type: string
          format: uri-reference
        hasLyrics:
          type: boolean
        metadataStatus:
          type: string
          enum: [LOCAL, SCRAPED, NEEDS_REVIEW]
        poolType:
          type: string
          enum: [PENDING, NORMAL, DISCOVERY]
        audienceType:
          type: string
          enum: [GENERAL, CHILD]
        genre:
          type: string
        region:
          type: string
          enum: [CN, KR, US, JP, OTHER]
        artists:
          type: array
          description: 艺人归属列表；不切割联名名称，包含“林俊杰”的名称统一归属为“林俊杰”
          items:
            type: string
...
    TrackPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Track"
        nextCursor:
          type: [string, "null"]
    Lyrics:
      type: object
      properties:
        plain:
          type: [string, "null"]
        synced:
          type: [string, "null"]
        source:
          type: [string, "null"]
...
    ScanStatus:
      type: object
      required: [state, discovered, imported, updated, skipped, failed, errors]
      properties:
        state:
          type: string
          enum: [IDLE, RUNNING, COMPLETED, FAILED]
        discovered:
          type: integer
        imported:
          type: integer
        updated:
          type: integer
        skipped:
          type: integer
        failed:
          type: integer
        message:
          type: [string, "null"]
        errors:
          type: array
          items: {type: string}
...
    Problem:
      type: object
      required: [status, title]
      properties:
        status:
          type: integer
        title:
          type: string
        detail:
          type: string